Changed lib.rs function
Reverted back to simple string in/output
