+++ formatter/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2024"
 
 [dependencies]
 lexer = { path = "../compiler/lexer" }
-source = { path = "../compiler/source" }
+
