+++ formatter/src/formatter.rs
@@ -3,28 +3,10 @@
 //! Processes tokens and produces formatted output.
 
 use crate::indent::IndentTracker;
-use crate::reorder::{categorize_keyword, parse_top_level_items, reorder_items, ItemCategory};
+use crate::reorder::{categorize_keyword, ItemCategory};
 use crate::spacing::requires_space;
-use lexer::{lex, Token, TokenType};
+use lexer::{Token, TokenType};
 use std::borrow::Cow;
-
-/// Formats Wasome source code and returns the formatted string.
-pub fn format_source(input: &str) -> String {
-    let tokens: Vec<Token> = lex(input).filter_map(|r| r.ok()).collect();
-
-    if tokens.is_empty() {
-        return String::new();
-    }
-
-    // Parse and reorder top-level items
-    let items = parse_top_level_items(tokens);
-    let reordered = reorder_items(items);
-
-    // Flatten back to tokens
-    let tokens: Vec<Token> = reordered.into_iter().flat_map(|item| item.tokens).collect();
-
-    format_tokens(&tokens)
-}
 
 /// Checks if a token starts a top-level item.
 fn is_top_level_start(token: &TokenType) -> bool {
@@ -192,7 +174,7 @@ impl TokenFormatter {
 }
 
 /// Formats a sequence of tokens.
-fn format_tokens(tokens: &[Token]) -> String {
+pub fn format_tokens(tokens: &[Token]) -> String {
     TokenFormatter::new().format(tokens)
 }
